requirements.txt
Pillow
pytesseract
tesseract
opencv-python
rembg